Several edits from experimentation with marshmallow_sqlalchemy. Basic CRUD ops for Inventory table curretly working.

+++ LabManager/equipments/routes.py
@@ -1,3 +1,4 @@
+# import marshmallow
 from flask import Blueprint, request, jsonify, render_template
 from flask_login import login_required
 from LabManager import db
@@ -17,49 +18,71 @@ def inventory():
 @equipments.route("/inventory/all", methods=["GET"])
 def inventory_all():
     inventory = Inventory.query.all()
-    result = equipments_schema.dump(inventory)
+    result = equipments_schema.dump(inventory).data
 
-    return jsonify(result.data)
+    return jsonify(result)
 
 
-@equipments.route("/inventory/<id>", methods=["GET"])
+@equipments.route("/inventory/<int:id>", methods=["GET"])
 def inventory_fetch(id):
-    equipment = Inventory.query.get(id)
-    result = equipment_schema.dump(equipment)
+    equipment = Inventory.query.get(id).first()
+    if equipment is None:
+        response = {
+                 'message': 'Inventory item does not exist'
+                   }
+        return jsonify(response), 404
 
-    return jsonify(result.data)
+    result = equipment_schema.dump(equipment).data
+
+    return jsonify(result)
 
 
 @equipments.route("/inventory/add", methods=["POST"])
 def inventory_add():
+    # new_data = request.json
+    # load = equipment_schema.load(new_data)
     name = request.json["name"]
     description = request.json["description"]
     new_equip = Inventory(name=name, description=description)
     db.session.add(new_equip)
     db.session.commit()
 
-    return equipment_schema.jsonify(new_equip)
+    # equipment_schema = new_equip.__marshmallow__()
+
+    return jsonify(equipment_schema.dump(new_equip).data)
 
 
-@equipments.route("/inventory/update/<id>", methods=["PUT"])
+@equipments.route("/inventory/update/<int:id>", methods=["PUT"])
 def inventory_put(id):
     equipment = Inventory.query.get(id)
+    if equipment is None:
+        response = {
+                 'message': 'Inventory item does not exist'
+                   }
+        return jsonify(response), 404
+
     name = request.json["name"]
     description = request.json["description"]
     equipment.name = name
     equipment.description = description
     db.session.commit()
 
-    return equipment_schema.jsonify(equipment)
+    return jsonify(equipment_schema.dump(equipment).data) # equipment_schema.jsonify(equipment)
 
 
-@equipments.route("/inventory/delete/<id>", methods=["DELETE"])
+@equipments.route("/inventory/delete/<int:id>", methods=["DELETE"])
 def inventory_del(id):
     equipment = Inventory.query.get(id)
+    if equipment is None:
+        response = {
+                 'message': 'Inventory item does not exist'
+                   }
+        return jsonify(response), 404
+
     db.session.delete(equipment)
     db.session.commit()
 
-    return equipment_schema.jsonify(equipment)
+    return jsonify(equipment_schema.dump(equipment).data)
 
 
 @equipments.route("/lendings")
